+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Fix deprecated data frame applymap warning
+
 ## 2.0.0 - 2026-01-07
 ### Added
 - Add Python 3.14 support