+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix deprecated data frame applymap warning
+- Fix complexity report link
 
 ## 2.0.0 - 2026-01-07
 ### Added