+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Removed
+- Remove unnecessary KeyboardInterrupt handling
+
 ## 2.1.0 - 2026-01-09
 ### Added
 - Add more unit and integration tests for higher coverage