+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Upgrade PieMaker to 2.2.0
+
 ### Removed
 - Remove unnecessary KeyboardInterrupt handling